Break down gamestate to smaller files
